+++ src/Controller/Admin/ImagesController.php
@@ -60,7 +60,7 @@ class ImagesController extends AdminController
 
         if ($request->isPost()) {
             $file = $request->getUploadedFiles()['file'];
-            $filename = $this->findAvailableFilename($file->getClientFilename());
+            $filename = $file->getClientFilename();
 
             $imagick = new Imagick;
             $imagick->readImageBlob($file->getStream());
@@ -146,26 +146,5 @@ class ImagesController extends AdminController
         return $this->renderPage($request, $response, 'admin/images/edit.html', [
             'image' => $image
         ]);
-    }
-
-    /**
-     * Find the next available filename if there are duplicates.
-     *
-     * @param string $filename Filename.
-     * @return string
-     */
-    private function findAvailableFilename(string $filename): string
-    {
-        for ($i = 0; $i < 100; $i++) {
-            $file = $this->ci->get('db')->getRepository('Kuusamo\Vle\Entity\Image')->findOneBy(['filename' => $filename]);
-
-            if ($file === null) {
-                return $filename;
-            }
-
-            $filename = FileUtils::increment($filename);
-        }
-
-        throw new Exception('Filename is already in use');
     }
 }